requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
us
